Move FPS and enemy cound from enemy to debug module
